+++ src/devpilot_core/identity/__init__.py
@@ -1,5 +1,6 @@
 from .models import DEFAULT_IDENTITY_REGISTRY_PATH, DEFAULT_IDENTITY_SCHEMA_PATH, LocalActor, LocalRole, permission_for_action
 from .rbac import IdentityRegistry, IdentityRegistryOptions, RbacCheckInput
+from .exposure import RbacExposureOptions, RbacExposureReporter
 
 __all__ = [
     "DEFAULT_IDENTITY_REGISTRY_PATH",
@@ -8,6 +9,8 @@ __all__ = [
     "IdentityRegistryOptions",
     "LocalActor",
     "LocalRole",
+    "RbacExposureOptions",
+    "RbacExposureReporter",
     "RbacCheckInput",
     "permission_for_action",
 ]